AC_PosControl: change int32/int16 to float in accel_to_throttle

+++ libraries/AC_AttitudeControl/AC_PosControl.cpp
@@ -346,7 +346,7 @@ void AC_PosControl::rate_to_accel_z()
 void AC_PosControl::accel_to_throttle(float accel_target_z)
 {
     float z_accel_meas;         // actual acceleration
-    int32_t p,i,d;              // used to capture pid values for logging
+    float p,i,d;              // used to capture pid values for logging
 
     // Calculate Earth Frame Z acceleration
     z_accel_meas = -(_ahrs.get_accel_ef_blended().z + GRAVITY_MSS) * 100.0f;
@@ -380,7 +380,7 @@ void AC_PosControl::accel_to_throttle(float accel_target_z)
     // get d term
     d = _pid_accel_z.get_d();
 
-    int16_t thr_out = (int16_t)p+i+d+_throttle_hover;
+    float thr_out = p+i+d+_throttle_hover;
 
     // send throttle to attitude controller with angle boost
     _attitude_control.set_throttle_out(thr_out, true);
